Restructured repo, added handpicked envs
