+++ src/distance/shaping.py
